+++ src/curvesimilarities/util.py
@@ -7,6 +7,7 @@ from scipy.spatial.distance import cdist
 
 __all__ = [
     "parameter_space",
+    "curvespace_path",
     "sample_polyline",
     "refine_polyline",
 ]
@@ -103,6 +104,46 @@ def parameter_space(P, Q, p_num, q_num):
     q_coord = np.linspace(0, q_vert[-1], q_num)
     Q_pts = sample_polyline(Q, q_coord)
     return cdist(P_pts, Q_pts), p_coord, q_coord, p_vert, q_vert
+
+
+def curvespace_path(P, Q, path, sample_num):
+    """Return point pairs in curve space defined by a path in parameter space.
+
+    Parameters
+    ----------
+    P : array_like
+        A :math:`p` by :math:`n` array of :math:`p` vertices in an
+        :math:`n`-dimensional space.
+    Q : array_like
+        A :math:`q` by :math:`n` array of :math:`q` vertices in an
+        :math:`n`-dimensional space.
+    path : ndarray
+        A :math:`N` by :math:`2` array of :math:`N` vertices of polyline in
+        curve space.
+    sample_num : int
+        Number of sample points to be uniformly taken from `path`.
+
+    Returns
+    -------
+    ndarray
+        A :math:`n` by :math:`2` by `sample_num` array of point pairs in curve space.
+
+    Examples
+    --------
+    >>> P = [[0, 0], [2, 2], [4, 2], [4, 4], [2, 1], [5, 1], [7, 2]]
+    >>> Q = [[2, 0], [1, 3], [5, 3], [5, 2], [7, 3]]
+    >>> _, path = ifd_owp(P, Q, 0.1)
+    >>> pairs = curvespace_path(P, Q, path, 100)
+    >>> import matplotlib.pyplot as plt  # doctest: +SKIP
+    >>> plt.plot(*np.array(P).T)  # doctest: +SKIP
+    >>> plt.plot(*np.array(Q).T)  # doctest: +SKIP
+    >>> plt.plot(*pairs, "--", color="gray")  # doctest: +SKIP
+    """
+    path_len = np.sum(np.linalg.norm(np.diff(path, axis=0), axis=-1))
+    path_pts = sample_polyline(path, np.linspace(0, path_len, sample_num))
+    P_pts = sample_polyline(P, path_pts[:, 0])
+    Q_pts = sample_polyline(Q, path_pts[:, 1])
+    return np.stack([P_pts, Q_pts]).transpose(2, 0, 1)
 
 
 def sample_polyline(vert, param):